version 3/requirements.txt
pandas>=1.5.0
pyarrow>=10.0.0
pymysql>=1.0.2
pyyaml>=6.0
paramiko>=2.11.0,<3.0
sshtunnel>=0.4.0
google-cloud-bigquery>=3.0.0
requests>=2.28.0
python-telegram-bot>=20.0